+++ src/main/java/com/kero/weatherstats/model/Strategy.java
@@ -3,5 +3,5 @@ package com.kero.weatherstats.model;
 import java.util.ArrayList;
 // Commit drillede lidt så Silas skulle ordne det for mig, ellers så har jeg lavet det (Adman).
 public interface Strategy {
-    double calculate(ArrayList<Integer>list);
+    double calculate(ArrayList<Double>list);
 }
